app/Http/Requests/Manager/Chapterディレクトリが肥大化しているための整理(yuki)
