+++ app/Http/Controllers/Api/Manager/ChapterController.php
@@ -4,16 +4,16 @@ namespace App\Http\Controllers\Api\Manager;
 
 use App\Exceptions\ValidationErrorException;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Manager\ChapterBulkDeleteRequest;
-use App\Http\Requests\Manager\ChapterDeleteAllRequest;
-use App\Http\Requests\Manager\ChapterDeleteRequest;
-use App\Http\Requests\Manager\ChapterPatchRequest;
-use App\Http\Requests\Manager\ChapterPatchStatusRequest;
-use App\Http\Requests\Manager\ChapterPutStatusRequest;
-use App\Http\Requests\Manager\ChapterShowRequest;
-use App\Http\Requests\Manager\ChapterSortRequest;
-use App\Http\Requests\Manager\ChaptersPatchStatusRequest;
-use App\Http\Requests\Manager\ChapterStoreRequest;
+use App\Http\Requests\Manager\Chapter\BulkDeleteRequest;
+use App\Http\Requests\Manager\Chapter\DeleteAllRequest;
+use App\Http\Requests\Manager\Chapter\DeleteRequest;
+use App\Http\Requests\Manager\Chapter\PutRequest;
+use App\Http\Requests\Manager\Chapter\UpdateStatusRequest;
+use App\Http\Requests\Manager\Chapter\PutStatusRequest;
+use App\Http\Requests\Manager\Chapter\ShowRequest;
+use App\Http\Requests\Manager\Chapter\SortRequest;
+use App\Http\Requests\Manager\Chapter\PatchStatusRequest;
+use App\Http\Requests\Manager\Chapter\StoreRequest;
 use App\Http\Resources\Manager\ChapterShowResource;
 use App\Model\Chapter;
 use App\Model\Course;
@@ -34,7 +34,7 @@ class ChapterController extends Controller
      *
      * @return ChapterShowResource|JsonResponse
      */
-    public function show(ChapterShowRequest $request)
+    public function show(ShowRequest $request)
     {
         // ログイン中の講師IDを取得
         $managerId = Auth::guard('instructor')->user()->id;
@@ -64,7 +64,7 @@ class ChapterController extends Controller
      *
      * @return JsonResponse
      */
-    public function store(ChapterStoreRequest $request)
+    public function store(StoreRequest $request)
     {
         // ログイン中の講師IDを取得
         $managerId = Auth::guard('instructor')->user()->id;
@@ -108,7 +108,7 @@ class ChapterController extends Controller
      *
      * @return JsonResponse
      */
-    public function put(ChapterPatchRequest $request)
+    public function put(putRequest $request)
     {
         // ログイン中の講師IDを取得
         $managerId = Auth::guard('instructor')->user()->id;
@@ -146,7 +146,7 @@ class ChapterController extends Controller
      *
      * @return JsonResponse
      */
-    public function delete(ChapterDeleteRequest $request)
+    public function delete(DeleteRequest $request)
     {
         // ログイン中の講師IDを取得
         $managerId = Auth::guard('instructor')->user()->id;
@@ -192,7 +192,7 @@ class ChapterController extends Controller
      *
      * @return \Illuminate\Http\JsonResponse
      */
-    public function bulkDelete(ChapterBulkDeleteRequest $request)
+    public function bulkDelete(BulkDeleteRequest $request)
     {
         // ログイン中の講師IDを取得
         $managerId = Auth::guard('instructor')->user()->id;
@@ -241,7 +241,7 @@ class ChapterController extends Controller
      *
      * @return JsonResponse
      */
-    public function deleteAll(ChapterDeleteAllRequest $request)
+    public function deleteAll(DeleteAllRequest $request)
     {
         // ログイン中の講師IDを取得
         $managerId = Auth::guard('instructor')->user()->id;
@@ -296,7 +296,7 @@ class ChapterController extends Controller
      *
      * @return JsonResponse
      */
-    public function sort(ChapterSortRequest $request)
+    public function sort(SortRequest $request)
     {
         // ログイン中の講師IDを取得
         $managerId = Auth::guard('instructor')->user()->id;
@@ -347,7 +347,7 @@ class ChapterController extends Controller
      *
      * @return JsonResponse
      */
-    public function updateStatus(ChapterPatchStatusRequest $request)
+    public function updateStatus(UpdateStatusRequest $request)
     {
         // ログイン中の講師IDを取得
         $managerId = Auth::guard('instructor')->user()->id;
@@ -385,7 +385,7 @@ class ChapterController extends Controller
      *
      * @return JsonResponse
      */
-    public function putStatus(ChapterPutStatusRequest $request)
+    public function putStatus(PutStatusRequest $request)
     {
         // ログイン中の講師IDを取得
         $managerId = Auth::guard('instructor')->user()->id;
@@ -418,7 +418,7 @@ class ChapterController extends Controller
     /**
      * 選択済みチャプターを公開/非公開にするAPI
      */
-    public function patchStatus(ChaptersPatchStatusRequest $request): JsonResponse
+    public function patchStatus(PatchStatusRequest $request): JsonResponse
     {
         // ログイン中の講師IDを取得
         $managerId = Auth::guard('instructor')->user()->id;
