+++ app/Http/Requests/Manager/Chapter/BulkDeleteRequest.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace App\Http\Requests\Manager\Chapter;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class BulkDeleteRequest extends FormRequest
+{
+    public function authorize()
+    {
+        return true;
+    }
+
+    public function rules()
+    {
+        return [
+            'course_id' => ['required', 'integer', 'exists:courses,id,deleted_at,NULL'],
+            'chapters' => ['required', 'array'],
+            'chapters.*' => ['required', 'integer', 'exists:chapters,id,deleted_at,NULL'],
+        ];
+    }
+
+    protected function prepareForValidation()
+    {
+        $this->merge([
+            'course_id' => $this->route('course_id'),
+        ]);
+    }
+}
