+++ app/Http/Requests/Manager/Chapter/DeleteAllRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Manager;
+namespace App\Http\Requests\Manager\Chapter;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class ChapterStoreRequest extends FormRequest
+class DeleteAllRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -24,7 +24,6 @@ class ChapterStoreRequest extends FormRequest
     public function rules()
     {
         return [
-            'title' => ['required'],
             'course_id' => ['required', 'integer', 'exists:courses,id,deleted_at,NULL'],
         ];
     }
